Major rewriting of the internals to use inheritance and method importing. This
leads to much smaller individual translation classes and hopefully will be
easier to support with less duplication of code. Note that the public function
interface is unchanged but the public access to variables has been removed
(and replaced with some class methods). From the user perspective the
translations are unchanged (some are known to be broken but will be fixed in a
future commit).

svn path=/trunk/perlmods/; revision=8782