HdrTrans: Provide a function for stripping a prefix
If a prefix is used in header translation, individual translation
methods back to FITS do not function unless the prefix is stripped.
This patch provides a clean_prefix() function that can be used
prior to calling determine_class() and the specific translation
method.

The alternative was to store the prefix inside the translated
hash, and modify each translation method to allow it to strip
automatically before processing. This would have been a big
patch and a processing overhead for something that is quite
rare.