acsis: Quick kluge to get things working with no SpecFrame
Some broken observations have CHANNEL frame rather than SpecFrame
so we have to protect against errors in getting or setting SpecFrame
attributes.

Main problem with this patch is that standard out gets an error
message about the lack of WCS since we try to get the value in
an eval rather than being clever about testing for the SpecFrame
before we try it. Do this because things are not working at the
telescope and they need a quick fix.