Rewrite calc_coords to use subheaders and to look at end AZEL as well as start
In some cases when we are translating a header from multiple data files
we need to look in subheaders. This should also work for older
Astro::FITS::Header modules that treat UNDEF as a special card for
merging.

Some observations miss the start AZEL so we need to be able to use the
end time and azel to calculate the rough RA/Dec position.