Document and augment the reverse-case in can_translate method.
In the end there was little wrong.  It had appeared that UTDATE was
being tested in the from-FITS direction, when it was in a strange
format: day of week, month day, year.  In fact it was the commentary
that was unclear.  So I reinvented the tests and only realised that I
had done so upon preparing this commit.