IRCAM: add rotation.
Rotation is necessary to mosaic IRCAM images correctly.
Therefore pre-reconfiguration data were retrieved from the
Wayback Machine:

http://web.archive.org/web/19970725044029/http://www.jach.hawaii.edu/~caa/ircam3_scale.txt

A date-dependant to_ROTATION function has been added
for IRCAM, with only the historical value of 1.5 degrees
present so far.  Other values should be entered when
their sense can be checked.

The content of the retrieved note is as follows:

ircam3 pixel scale and image orientation

On AIC night 960209 I (finally) re-quantified the ircam3 pixel scale at
J, H and K and also the array orientation wrt n-s.  I put a star in two
corners of the array (and noted the telescope offsets betwen them). Here
are the results:

Pixel scale :    J            H            K
           0.2846"/pix  0.2839"/pix  0.2836"/pix

The nominal value of 0.286" is therefore correct to 2 parts in a 1000.

The array orientation is, as we have suspected in the past from mosaic
tiling errors, not n-s there is a tile in the array of -1.5 degrees wrt
n-s with n being 1.5 degrees east of north.  In ircamdr there is a
command rotoff that will rotate a set of telescope offsets by a
user-defined amount (in degrees).  For mosaics the user should rotate
his/her offsets by -1.5 degrees (1.5 degrees counter-clockwise) and use
that to mosaic images together.  Test mosaics taken on the above night
mosaiced together excellently with the rotated telescope offsets.