ACSIS: Add custom translation of TRANSITION
This is to tidy up the whitespace in ACSIS TRANSITI headers.  The reason
for this is that we might like to tidy up the whitespace in the transition
names in the OT's line catalog at some point, so this translation will allow
us to set up, for example recpars filters, with the tidied transition names
in advance and not need to change them if and when the line catalog is updated.

The reverse translation currently restores the double space before the
dash so that the untranslated header matches the original.