ACSIS: Debug the _REFERENCE_LOCATION translation.
Correct the sprintf formatting to give two decimal places not two
digits.

Due to some weird Perl setup on my server, I wasn't able to debug this
originally.  In practice I was able to test the ORAC-DR code it was
intended for without it.  After a couple of weeks rather than lose the
modification I committed it as nobody else would be using it.  Today I
had a reduction that incorporated multiple reference positions that
needed independent analysis, forcing me to re-open this pending issue.