JCMT: Add translation of dome status headers
JCMT data files now include 4 headers indicating the status of the
dome: DOORSTST, DOORSTEN, ROOFSTST and ROOFSTEN.  These can either
be "Open" or "Closed" (or undefined).

For convenience of checking the dome status, they are combined into
a single "DOME_OPEN" generic header.  This is true if all 4 of the
headers say "Open", false if all of them say "Closed" and undefined
otherwise.  (The disadvantage of this approach is that, when converting
back to FITS headers, information is lost in the case of mixed
values.)