JCMT: check opacity values are not zero
Now when looking for opacity values, as well as checking for
defined values, also check that they are not zero.

This retains the current logic (prefer CSO to JCMT WVM, but only
if both start and end values are present).  It might be better
to add a date beyond which CSO values are considered unreliable
and/or to compare the associated timestamps (e.g. WVMDATST) with
the date of the observation to ensure they are close enough to be
relevant.